Added Conclave Code
Connection Handler and conclave.start(). The issue seems to be that when I connect using the client code at the bottom, it will only read the first connection's message and handle it, but a second connection i make using the same client code does not get written to the system and I always get a server error.